perf: Use Cow for value in MergedCSTNode (#64)
In most scenarios, we don't need to create a new `String` when merging
the values of two CSTNodes. More often than not, we simply pick the
value from either the left or right node, or, ideally, the values of
both nodes are equal to the base. Consequently, creating owned Strings
is uncommon, and it would be more efficient to return a reference to the
value of one of the `CSTNodes`.

To accommodate this use case, this PR updates the `MergedCSTNode` to use
`Cow<str>` instead of `String`. This change eliminates unnecessary
`to_string` calls, potentially leading to performance improvements in
some scenarios.