perf: Remove expensive call to find matching in children (#66)
During the merge process, given two nodes, left and right, we need to
iterate over their children to determine how to merge them. In each
iteration, we try to find a match for the current left child among the
right children and vice versa. Previously, we were examining all
captured matches across the entire program tree, which was heavily
inefficient.

This PR changes the implementation to only consider the matches captured
in the other node's children, significantly reducing the number of
iterations required. Initial observations in empirical scenarios showed
a performance boost of nearly 3x.